Player Movement (in theory)
Implemented player movement for the QBert, although only in theory - the game still crashes once SDL is included, so the code to set up the input is temporarily commented out.

I first turned Pyramid into an Observer that would update the QBert Graphics Component when prompted, but I later realized the pyramid shouldn't be responsible for all the characters movements, and therefore created a Movement Component (also an Observer) that the QBert GameObject could be assigned with.
It was only after making this new component that I realized that I might want to implement different movement patterns depending on the character, so I scrapped that new class and just applied the movement code to the QBert component itself (that was still rather empty).