Block Color Change and Improved Sprites
Made the blocks change their color to yellow when stepped on. To do so, I created a MainObserver that receives a raw pointer to the Pyramid as an argument and keeps check of QBert - once QBert sends a notification that their position index has changed, MainObserver calls the TurnCube() function on the block with the respective index in the Pyramid (and once all blocks are yellow, a "level complete" function is called, although it still doesn't do much);

Changed both the GraphicsComponent and the Renderer classes as to be able to define the source rectangle of a sprite, and therefore only draw a portion of said sprite;

Changed the QBert image for a proper sprite sheet (using the new source rectangle definition feature), and made the drawn sprite change according to the direction they're facing.

(( The bug described in the last push is still not fixed, but as I don't plan to move any GameObject's position instead of its components in the near future, and I still have no TextComponents implemented (the only other class that uses Tranform), I've come to the conclusion fixing it right away shouldn't be my main priority. ))