All Level Maps and Cube-Step Mechanics
Changed the SceneManager as to only render and update the current scene and created a function to switch between scenes;

Changed the Cube image to a spritesheet, and changed both its class and Pyramid's to allow the specification of Level and Color in the constructors. Now the cubes know to behave differently when stepped on depending on the level, as well as in which color pattern to be drawn depending on the level section;

Renamed the MainObserver class to LevelSectionObserver for readability purposes;

Created all the scenes, 4 for each level, and made them change automatically once all the blocks are turned (through the LevelSectionObserver class);

Wrote a function to create the QBert GameObject only once and storing it as a global variable (and also turned some other variables that should always be the same, no matter the level, into globals). This way, I use the same QBert in all the scenes, avoiding having to create a new one every time, as well as reset all the player input;

And tried to create a GameObserver class that would keep track of all LevelSectionObservers and print out a message every time a level is completed, but I didn't manage to get it to work. I still don't understand exactly how to make such a class functional, as an instance of it would have to be created before the first scene's initialization, but somehow it should still be able to keep track of all the LevelSectionsComponents in the rest of all the not yet created scenes - which cannot be initialized before said scene as they need other level specific components and information.