Slick, Sam and QBert Position Reset
Made QBert's visuals reset between each scene. To do so, I had to add a function to QBertComponent that changes the QBert GraphicsComponent sourceRect and position, so that LevelSectionObserver doesn't need any QBert graphics specific information;

Created a SlickSam component that behaves similarly to the QBert one, but not only handles no health-related information, but also updates the position after an assignable (through the constructor arguments) amount of time, either randomly going diagonally left or right;

Created a RemoveAllComponents() function in the GameObject class, as to be able to clear a game object of all its functionalities once it becomes useless to the scene - and then used said function in the new SlickSam component whenever they die;

Made the SlickSam component a subject of the LevelSectionObserver, calling a specific event every movement as to switch the stepped-on cubes colors appropriately. To do so, I had to add a second constructor to the observer, allowing an extra argument of a vector of pointers to SlickSam components;

Made Slick and Sam die when touched by QBert, by making LevelSectionObserver check all characters position indexes and calling out the respective death functions when a collision's detected.