Factory Class and Slick/Sam Spawners
Created a Factory class to be able to spawn different game objects with less code repetition, and cleaned the project of most other game object creations;

Made the LevelSectionObserver spawn Slicks and Sams every couple of seconds if requested. For that, I had to remake the LevelSectionObserver constructor as to not take a vector of SlickSam components, but instead just take a bool specifying if it should spawn (or not) and a float for the spawn interval. If no spawning is intended, the SlickSam vector will stay empty, otherwise it will create new foes with the new Factory class and add them to the vector to be properly observed;

Edited every level section as to spawn Slicks and Sams at an increasing pace - no spawning in the 1st level, in the 2nd one spawning every 20-15-10-5 seconds (according to section), and spawning every 25-20-15-10 seconds in the 3rd one (considering Slick/Sam's abilities become way more difficult to deal with in the last level, once the cube activation rules change).