Ugg and Wrongway
Created a UggWrongway component class somewhat similar to SlickSam, but now moving only to the sides or up. Even though their movement looks bizarre visually, I still considered that whatever cube their standing on (on the side, that is) should still be their position index;

Made the LevelSectionClass able to spawn Uggs and Wrongways optionally and properly observe them. To do so, I now also store pointers to the created components in a vector, and whenever any Ugg/Wrongway moves (and the event is called) I check for collisions with the cube they're hovering over, not the one they stand on;

Fixed a bug that would lead my game to crash after a while duo to changing the enemies vectors while inside a for-loop going through said vectors contents;

Edited all level sections again as to spawn Uggs and Wrongways at an increasing pace - in the 1st level spawning  every 10-7 seconds, but only from section 3 onwards, and in the 2nd and 3rd levels spawning every 15-10-7-5 seconds.