New Level Screens
Added a scene between each level that lasts around 2 seconds and shows an image with the new level number. To do so, I wrote a new simple constructor for the LevelSectionObserver which only takes the time the title screen should be shown (and, through the update, calls the SceneManager to skip the scene after said time). Aside from that, I also added a function to the Factory class to create this new simplified LevelSectionObserver - as to avoid repeating the same code every level set-up.