QBert's Death, Victory Screen and Bug Fixing
Made a Victory Screen to show once the player completes the 3rd level. At this point, it's just an empty scene with an image of the victory message and little trophies;

Made QBert be initialized frozen and get player input enabled back in the beginning of every section (and disabled again in the end). I realized that, without this, the player would still be able to move during the title screens, as the QBert game object is stored outside of the scenes;

Added a better visual queue for the QBert's death - now, whenever they get hurt, everything on the map gets frozen for 2 seconds while they curse. Afterwards, the map is emptied out and, after 1 second, QBert appears again in the cube they died in (with player input enabled back).