Transform Bug Fix and Jumping Observer Start
Fixed my Transform class bug (mentioned in previous push descriptions) and added an FPS counter in each level section to test it out;

Made QBert not be under the DAE namespace, as to separate what classes are originally from the engine and which ones are not;

Created an observer for the QBert component called JumpingObserver that will, in the future, produce an actual jumping animation when the player moves, instead of just teleporting the character. During the animation, the character can't prop any collision events - only after landing.
The animation part itself has not been coded yet, so it still just teleports when the event is prompted - but the movement is already executed in the new observer, not on the QBert class. This class is also still only applicable to QBert, but I intend on making it capable of observing other character components as soon as possible (by making different constructors.