Game UI and Bug Fixing
Made all the needed UI, adapted not only for the Single-Player Mode, but also Co-op Mode (which I still haven't developed) - meaning that all the display constructors take at least a boolean as an argument that specifies which mode is currently being played, and with which the class organized the position of the UI on screen accordingly;

Noticed a bug that crashed the program whenever more than one Ugg/Wrongway was spawned and fixed it accordingly. I didn't have this issue before and I couldn't and can't understand why it started now, but a reinforcement check in the scene's GameObjects update for-loop fixed it.