Coily Start and More Bug Fixing
Created a Coily component which, as of now, behaves similarly to Slick/Sam but damages on contact instead of just dying (and also has a different sprite while mid-jump). The class was set up as to act differently while Coily is an egg from when they're a snake (drawing different sprites and moving differently) and switch modes once the last pyramid row is reached, but because I haven't programmed the movement yet, they just jump off and die for now;

Made Coily spawn in every LevelSectionObserver 3 seconds after the beginning of the round, and respawn again 8 seconds after they die - and because every round has a Coily, no constructor arguments were added to the observer class;

Fixed 2 important bugs: one led to foes not changing sprites when switching direction for their final jump (at the edge of the pyramid), and the other led to foes not freezing once QBert died.