Flying Disk and Coily Adjustment
Created a Disk component class which brings a given QBert and itself up to the top of the pyramid once activated. The activation is called through a function that takes, as arguments, a QBert component and a GraphicsComponent (also from the QBert game object). The LevelSectionObserver takes care of calling said function when needed, as well as informing the disks when to delete themselves - but the disks are still created individually through a Factory function and referenced to said observer through its constructor (as a pointer to the vector that holds them);

Adjusted Coily's behaviour as to always jump off the map whenever chasing a QBert at a close range that also jumps off - because, considering Coily gets frozen and reset if QBert dies, they will only keep moving if QBert gets on a disk.