Start Screen and Improved Victory Screen
Made a start menu which became the 1st initialized scene. To do so, I wrote a new component class named StartScreen that takes the indexes of the first scenes in each game mode (solo, co-op or versus), as to load the appropriate mode depending on the player selection;

Changed the victory screen to not only show the score you finish with, but also allow you to go back to the start menu.