Fixed Scene Reload and UI
Made every scene able to be played more than once - as to allow finishing the game, going to the start menu and playing the same mode again. For that to work, I had to add a "Reset" function to some components, make the QBert's round, level, lives and score reset once the MenuScoreDisplay is drawn (as it is only drawn if the player dies or wins the game) and change the LevelSectionObserver class as to reset all its members during the Initialize() function and not delete any disks from its vector, only hide them;

Changed how the UI is created and drawn, so that no display component is stored in QBert, and so that the component classes that take care of  the displays logic don't have to also create the visuals and add them to the scene. Now said components constructors have to take the GraphicsComponents or TextComponents they intend on working with;

Changed where the QBert's falling sound is played, as before it was always called whenever QBert jumped off the map - even if they had taken a disk.