Player Input Components
Made all player input be set up through newly written components (named SetPlayerOneControllers and SetMenuControllers, for now), so that they're defined only when necessary - therefore stopping the menu controllers from working during the game, and vice-versa. For that to work, I also created a function in the InputManager class named ClearAllCommands(), which gets called whenever the inputs need to change. As the name entails, this new function gets rid of all previous defined input in the manager.