Death Screen and Bug Fixing
Created a Death Screen scene, functionality-wise similar to the Victory Screen one, and changed to it whenever the player loses all their lives (after the death animation, that is);

Changed some code order in the Input Manager, so that when a command is called, its "executed" boolean is set to true right before actually executing said command, instead of after (like I had before). These bools are the ones used to make the command act according to their "press type" ("while pressed", "when pressed", "when released", etc) and their alteration being before or after the actual execution is irrelevant for the class - but it did create issues before in the cases that a command was used for a scene transition, as that would sometimes entail the deletion of all previous commands;

Fixed a bug that would lead to the hearts in the UI not reappearing on reloaded levels.