Co-op Death and Victory Screens
Added mode-specific death and victory screens for the Co-op Mode (as I had done to the other modes before). To do so, I had to create three new Factory classes - one to make the death visuals, one to make the victory visuals, and one to deal with the logic of the two MenuScoreDisplays in both screens (unlike the logic in the other modes, which only deals with one score).