Remaining Foes State Machines
Changed both remaining enemies (Slick/Sam and Ugg/Wrongway) update functions, as to apply the State Machine design;

Added a little sound cue to the StartMenu that plays whenever the player changes selection;

And revised all the mode's Level 02 setup files, as to make the Ugg/Wrongway spawn timer 13 secs instead of 15 in the 1st round. I did this so they won't spawn in complete synch with the Slick/Sams.