LevelSectionObserver State Machine and Bug Fixing
Made LevelSectionObserver's update function work with a StateMachine;

Fixed a couple of bugs in LevelSectionObserver, most of them related to Co-op Mode and the fact a player could still die or win the game after their companion had died (as long as they'd jump right before their companion's death). This would lead to bizarre behaviour, like the game continuing after a player had lost their last life (and their sprite just staying frozen wherever they last died), players respawning still outside the map after a fall, etc. To fix all this, I just made it so that after a QBert's death or a victory, any movements still being executed will have no effect (to give some examples, no cubes will be turned in a jump that lands after a death and no damage will be given for a foe collision after a victory). The only exception to this rule is the graphics repositioning of a player that falls after a prior death/victory - in those cases, the fallen QBert will revert to the last cube they've stood on, although still not taking damage;

Lastly, changed the way the Lives Display hides the hearts representative of a player's lost lives - has it was causing some issues if the player did more than two runs in a row without closing the game. Now, the LivesDisplay constructor takes the original X position of the drawn hearts as an argument, as to not use unreliable and unprofessional fixed values (like before) - and said issues got fixed.