Coily's Bug Fixing
Added two RemoveCommand() functions to InputManager, as to be able to remove a specific command binding without disrupting the rest of the map;

With these new functions, fixed a bug I noticed yesterday regarding the P2 Coily input - if P2 would press any key in the time between Coily's death and their respawn, the game would crash. This would happen because the commands would try to access a deleted component so, to fix this, I turned P2CoilyInput into an observer and every time a Coily dies, all the P2 controls are removed. The deletion of the components (aka, calling Coily's Die() function) is now also taken care of by this new observer instead of the LevelSectionObserver, because a component can only be deleted as a result of a subject notification if said deletion happens in the last added observer;

Fixed another tiny bug that would make QBert take damage prematurely if moving under a Coily still in the middle of their spawning fall. This would happen due to there being no check for Coily's m_Airborne bool variable when calculating their collision.