Color Display, Code Clean-Up and More Bug Fixing
Created a Color Display to make it clearer to the player which color should the cubes be turned into in each round. To do so, I also had to change the UI layout a bit;

Changed the RoundLevelDisplay class as to not inherent from Observer, nor depend on any QBert variables, since said display doesn't need to update at runtime and can just take the original round and level ints as constructor arguments. This also led me to stop storing game-related variables in Qbert;

Discovered that, in Co-op Mode, if both players jumped in a flying disk at the same time, the last one to land would not die and get frozen after the jump. This happened due to the LevelSectionObserver not checking the disks bool variable "m_Activated" on collision, and only checking m_HasBeenUsed. While the second one returns false until the flight animation is complete, m_Activated will only return true while the animation's running.