Pause Screen Fixes and Disks Spawn Improvement
Realized the Pause Screen was not drawn over any foes, as said foes were spawned and added to the scene after the screen's GameObject. To fix this issue, I split the container m_Objects in the Scene class into 2 containers - one for objects with rendering priority and one for normal objects. I also changed the Add() function as to allow for the specification of the object's render priority through a non-obligatory bool (defaulted to false) when adding it to the scene. With everything neatly split, the Render() function now renders the normal container first and only afterwards the one with priority;

Also noticed that the disks would not change their original position on round reset if the player didn't close the game, as they were created during set-up and only given as a reference in the arguments of the LevelSectionObserver constructor. To fix this, I made LevelSectionObserver spawn the needed Disks by itself during initialization, therefore spawning new disks (with new randomized positions) every round reload;

Fixed a couple bugs related to the pausing mechanic and the flying disks (for example, they'd stay paused forever if a QBert died or a player went back to the Main Menu without losing/winning the game).