Instructions Update and General Adjustments
Edited the Instructions Screens visuals as to also let the player know about the pause buttons;

Added controller input for Player 2 in the Main Menu screen;

Fixed a bug that would allow any foe to collide with the second player's QBert during Solo Mode if they reached the top cube. This would happen because both QBerts are always given to the LevelSectionObserver and, depending on the mode, the observer will take in account the appropriate Qbert/s - but I was still checking collisions with both players whenever any entity moved.