Final Push
Deleted all useless code I found throughout my project, added more comments as to make everything more readable, and generally tried to improve the organization of my classes;

Made all classes that draw textures actually stop drawing them when hidden, not just draw them outside the screen like I've done up until now (for example, with the pause screen or the UI hearts). I did so by adding a new member variable and a respective setter to GraphicsComponent called m_Hidden, which is checked during the Render() function;

Fixed Player 2's controller input (both as Coily and as Qbert), as pausing with the start button in a second controller as intended didn't work. Also changed  the controller button to go back to the main menu (be it in a pause, death or victory screen) to Y, as to not have different inputs for the same action at different points in the game;

Split the disk flight sound into two files - the ascension and the QBert fall. With them separated, now I can execute the falling one only when the disk disappears, instead of relying on the animation time and the sound time matching perfectly. This also made the sound way less jarring if the player pauses the game right after jumping into a disk;

Centred the level/round display in Co-op Mode, because it ended up with a tiny offset when I moved the pyramid a couple days ago;

Fixed the post-build events in the game project as to output executables that run properly. Specifically, added a command to copy the level .bin files into each build-specific sub-folder, as well as the Data folder and all its contents into both platform folders.